Renamed some variables, organized some styles
